Fix inconsequential yet erroneous slip-up
The commit object name may be passed as a third argument to that hook in
some cases (such as amend commits).  I should note that they may not be
always SHA-1 any more, but in practice, they still are, and will likely
remain as such for a fair while, see:

    https://git-scm.com/docs/githooks#_prepare_commit_msg

In any case, this was only the remnant traces of some debugging that
occurred during a recent polishing session of that script.

Ref: 10539fa