Forget about gpg-encoding filters and its content
It was such a pain to work with when rebasing.  I spent about 16 hours
over the week-end to try git-crypt, which doesn't play so well with
dirty submodules; with symmetric and/or deterministic encryption, etc...

Ultimately, each copy would contain *all* the data again—how wasteful
from the perspective of versioning.  In the end, I do not actually care
at all either about what my time-logs looked like at a particular time,
they're not exactly code that may be reverted savantly, but rather data
whose latest state is what it is.

I'll use Syncthing to share it across devices, just in case; and in the
future, will have some recurring systemd task that backs it up to some
of my servers.