Avoid the 1.0-proportion for window height
It's a fairly hacky implementation that would then, for example, cause
consumed windows to be stacked strictly underneath, and therefore
entirely out of view.  In the end, maximising, expelling, or somehow
else moving the window is a more reliable way to adequately restore its
natural trait for full vertical spanning.